Moved dummy problem to file
